# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Inference Intelligence Platform API
  version: 2.5.0
  description: Local-first, provider-agnostic multi-agent intelligence platform serving Trading Bot, FORGE, Nexus, Sentinel, and FRIDAY.
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.Inference.internal
    description: Production internal cluster
paths:
  /v1/trading/consult:
    post:
      summary: Trading Consultation Advisory
      description: Returns bounded trading parameter advisory for autonomous trading bots. Inference NEVER touches private keys or executes trades directly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol, current_regime]
              properties:
                symbol: { type: string, example: "BTC/USDT" }
                current_regime: { type: string, example: "high_volatility" }
                drawdown_pct: { type: number, example: 3.5 }
      responses:
        '200':
          description: Advisory parameter bounds
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendation: { type: string }
                  confidence: { type: number }
                  bounded_parameters: { type: object }
...